--- oneTrillion/src/project_02/GameStartScreenPanel.java
@@ -1,196 +1,0 @@
-package project_02;
-
-import java.awt.AlphaComposite;
-import java.awt.BasicStroke;
-import java.awt.Color;
-import java.awt.Cursor;
-import java.awt.Graphics;
-import java.awt.Graphics2D;
-import java.awt.RenderingHints;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
-
-import javax.swing.ImageIcon;
-import javax.swing.JButton;
-import javax.swing.JPanel;
-
-public class GameStartScreenPanel extends JPanel implements Runnable {
-	// 원의 시작 위치
-	private int x, y, Ox, Oy, r;
-	double th;
-	Thread thread;
-	private InsideOut insideOut;
-
-	// fadeIn과 fadeOut 을 위한 변수
-	private float fadeValue;
-	private boolean isFadeOut;
-
-	// MainScreen 제어를 위한 변수
-	private boolean isMainScreen;
-
-	// 뒤로가기 버튼 이미지를 담을 수 있는 객체
-	private ImageIcon backButtonImage = new ImageIcon(
-			getClass().getClassLoader().getResource("images/backButtonImage.png"));
-
-	// 마우스가 버튼에 진입했을 때의 이미지
-	private ImageIcon backButtonEnteredImage = new ImageIcon(
-			getClass().getClassLoader().getResource("images/backButtonEnteredImage.png"));
-
-	// JButton 구현
-	private JButton backButton = new JButton(backButtonImage);
-
-	GameStartScreenPanel(InsideOut insideout) {
-		this.insideOut = insideout;
-		// 쓰레드를 만들고 실행시켜준다.
-		setThread(new Thread(this));
-		// 컨테이너의 크기가 변경될때 컴포넌트들의 크기와 위치가 자동적으로 변경되는데 그걸 해제한다
-		setLayout(null);
-		// 게임창 크기 설정
-		setSize(Main.SCREENT_WIDTH, Main.SCREENT_HEIGHT);
-		setBounds(0, 0, Main.SCREENT_WIDTH, Main.SCREENT_HEIGHT);
-		// 검정색 바탕에 흰색 원 이므로 검정색으로 설정
-		setBackground(Color.BLACK);
-		// 화면 출력 설정 기본값은 false 이므로 설정 해줘야한다.
-		setVisible(true);
-		// backButton의 위치 설정
-		buttonSet(backButton, 80, 60, 228, 57);
-
-		/**
-		 * backButton의 마우스 이벤트를 처리해준다.
-		 */
-
-		backButton.addMouseListener(new MouseAdapter() {
-			/**
-			 * 마우스가 아이콘 위에 있을때 이벤트 처리
-			 */
-			@Override
-			public void mouseEntered(MouseEvent e) {
-				// 아이콘 이미지를 Entered 이미지로 변경
-				backButton.setIcon(backButtonEnteredImage);
-				// 커서 이미지도 HAND_CURSOR로 변경해서 좀더 알아보기 쉽게한다.
-				backButton.setCursor(new Cursor(Cursor.HAND_CURSOR));
-			}
-
-			// 마우스가 아이콘을 벗어 났을때 이벤트 처리
-			@Override
-			public void mouseExited(MouseEvent e) {
-				// 아이콘 이미지를 기본이미졸 변경
-				backButton.setIcon(backButtonImage);
-				backButton.setCursor(new Cursor(Cursor.DEFAULT_CURSOR));
-			}
-
-			// 마우스가 backButton 아이콘 눌렀을때 이벤트 처리
-			@Override
-			public void mousePressed(MouseEvent e) {
-				isFadeOut = true;
-				isMainScreen = true;
-			}
-		});
-
-	}
-
-	// run 함수에서 while문을 통해 계속 화면을 그려줌으로써 다음 화면으로 넘어갈 수 있게 해준다.
-	@Override
-	public void run() {
-	fadeIn();
-		while (true) {
-			repaint();
-			try {
-				th += 1.0;
-				this.x = (int) (r * Math.cos(th / 180 * Math.PI));
-				this.y = (int) (r * Math.sin(th / 180 * Math.PI));
-				if (isFadeOut && isMainScreen) {
-					fadeOut();
-					insideOut.changeMainScreen();
-					return;
-				}
-				Thread.sleep(10);
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		}
-
-	}
-
-	public void paintComponent(Graphics g) {
-		super.paintComponent(g);
-		// graphics를 2D로 변경
-		Graphics2D g2 = (Graphics2D) g;
-		// 투명도를 조절하기 위한 부분 fadeValue 가 1.0이면 불투명도 100%, 0.1이면 불투명도가 10% 이다.
-		g2.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_OVER, fadeValue));
-		// 안티엘리어싱
-		g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-		// 색깔 설정
-		g2.setColor(Color.WHITE);
-		// 두께설정
-		g2.setStroke(new BasicStroke(8));
-		// 원을 그리는 메소드 x좌표, y좌표, 오른쪽 왼쪽 지름(width), 위 아래 지름(height)
-		g2.drawOval(380, 100, 530, 530);
-		// drawOval을 통해서 원을 그리면 내부가 비어있는 원이 나오지만 fillOval을 통해 그리면 내부가 지정된 색으로 채워져 그려진다.
-		g2.fillOval(x, y, 25, 25);
-	} // 630 , 73
-	
-
-	/**
-	 * fadeIn 효과를 주기위한 메소드 temp를 사용한 이유는 fadeIn값이 1.0을 넘어가면 에러가 발생하기 때문에 float연산 특성상
-	 * 0.1씩 10번 증가시키면 1.0이 아니라 1.000001이 되서 에러가 발생한다. 따라서 temp를 증가시키고 fadeIn에 대입시키는
-	 * 방식을 사용한다. 여기서 temp가 1보다 커지면 temp를 1로 설정하고 대입시켜준다.
-	 */
-	public void fadeIn() {
-		try {
-			float temp = 0;
-			fadeValue = 0;
-			while (fadeValue < 1) {
-				temp += 0.1;
-				if (temp > 1) {
-					temp = 1.0f;
-				}
-				fadeValue = temp;
-				repaint();
-				Thread.sleep(100);
-			}
-		} catch (InterruptedException e) {
-			e.printStackTrace();
-		}
-	}
-
-	public void fadeOut() {
-		try {
-			float temp = 1.0f;
-			fadeValue = 1.0f;
-			while (fadeValue > 0) {
-				temp -= 0.1;
-				if (temp < 0) {
-					temp = 0;
-				}
-				fadeValue = temp;
-				repaint();
-				Thread.sleep(100);
-			}
-		} catch (InterruptedException e) {
-			e.printStackTrace();
-		}
-	}
-
-	// 일일이 다 설정하기 힘드므로 메소드를 통해 손쉽게 버튼의 위치를 설정
-	public void buttonSet(JButton button, int x, int y, int width, int height) {
-		button.setBounds(x, y, width, height);
-		// 버튼 테두리 제거
-		button.setBorderPainted(false);
-		// 누르는 느낌 제거
-		button.setContentAreaFilled(false);
-		// 글씨 테두리 제거
-		button.setFocusPainted(false);
-		// 버튼 추가
-		add(button);
-	}
-
-	public Thread getThread() {
-		return thread;
-	}
-
-	public void setThread(Thread thread) {
-		this.thread = thread;
-	}
-
-}
